@@ -58,33 +66,31 @@
     "webpack-md5-hash": "0.0.5"
   },
   "devDependencies": {
+    "babel-cli": "^6.10.1",
     "babel-eslint": "^6.0.4",
-    "babel-plugin-transform-flow-strip-types": "^6.8.0",
+    "babel-plugin-transform-flow-strip-types": "^6.7.0",
     "chai": "^3.5.0",
-    "chai-enzyme": "^0.4.2",
+    "chai-enzyme": "^0.5.0",
     "cheerio": "^0.20.0",
     "coveralls": "^2.11.9",
     "enzyme": "^2.2.0",
-    "eslint": "^2.8.0",
-    "eslint-config-airbnb": "^9.0.0",
-    "eslint-import-resolver-webpack": "^0.2.4",
+    "eslint": "^2.11.0",
+    "eslint-config-airbnb": "^9.0.1",
+    "eslint-import-resolver-webpack": "^0.3.0",
     "eslint-plugin-flow-vars": "^0.4.0",
     "eslint-plugin-import": "^1.6.1",
-    "eslint-plugin-jsx-a11y": "^1.2.0",
-    "eslint-plugin-react": "^5.0.1",
+    "eslint-plugin-jsx-a11y": "^1.3.0",
+    "eslint-plugin-react": "^5.2.2",
     "eventsource-polyfill": "^0.9.6",
-    "expect": "^1.19.0",
-    "expect-jsx": "^2.5.1",
-    "flow-bin": "^0.25.0",
+    "flow-bin": "^0.27.0",
     "iflow-immutable": "^1.0.32",
-    "iflow-react-redux": "^1.0.15",
-    "iflow-react-router": "^1.0.42",
-    "iflow-react-widgets": "^1.0.16",
+    "iflow-lodash": "^1.1.14",
+    "iflow-react-redux": "^1.1.14",
+    "iflow-react-router": "^1.1.14",
     "iflow-redux": "^1.0.37",
     "iflow-redux-actions": "^1.0.3",
     "isparta": "^4.0.0",
     "isparta-loader": "^2.0.0",
-    "jscs": "^2.11.0",
     "karma": "^0.13.22",
     "karma-chrome-launcher": "^1.0.1",
     "karma-coverage": "^1.0.0",
