@@ -1,8 +1,10 @@
-/* eslint-disable global-require */
-
 /**
- * This is the entry file for the application.
+ * @file    application entry point for the client
+ * @author  Sebastian Siemssen <sebastian@amazeelabs.com>
+ * @date    2016-01-01
  */
+
+/* eslint-disable global-require */
 
 // Needed for redux-saga es6 generator support.
 import 'babel-polyfill';
@@ -16,11 +18,10 @@ import ReactDOM from 'react-dom';
 import Relay from 'react-relay';
 import { browserHistory } from 'react-router';
 import { syncHistoryWithStore } from 'react-router-redux';
-import withScroll from 'scroll-behavior';
 import configureStore from 'configureStore';
 
 // Set up relay.
-const environment = new Relay.Environment();
+const environment = new Relay.Environment();  // Set up relay.
 environment.injectNetworkLayer(new Relay.DefaultNetworkLayer('/graphql'));
 
 // Create redux store with history. This uses the singleton browserHistory
@@ -35,9 +36,9 @@ const store = configureStore(initialState, browserHistory);
 // non-default key ("routing"), selectLocationState must be provided for
 // resolving how to retrieve the "route" in the state
 import locationStateSelector from 'selectors/locationStateSelector';
-const history = withScroll(syncHistoryWithStore(browserHistory, store, {
+const history = syncHistoryWithStore(browserHistory, store, {
   selectLocationState: locationStateSelector,
-}));
+});
 
 // Find the DOM node generated by the server.
 const mountNode = document.getElementById('app');
@@ -50,7 +51,7 @@ if (module.hot) {
   const renderApp = render;
   const renderError = (error) => {
     const RedBox = require('redbox-react');
-    ReactDOM.render(<RedBox error={error} />, mountNode);
+    ReactDOM.render((<RedBox error={error} />), mountNode);
   };
 
   render = () => {
@@ -67,7 +68,8 @@ if (module.hot) {
 // Do the initial rendering.
 render();
 
-// Install ServiceWorker and AppCache in the end since it's not the most
-// important operation and if main code fails, we do not want it installed.
-import { install } from 'offline-plugin/runtime';
-install();
+if (__PRODUCTION__) {
+  // Install ServiceWorker and AppCache in the end since it's not the most
+  // important operation and if main code fails, we do not want it installed.
+  require('offline-plugin/runtime').install();
+}
