@@ -18,11 +18,9 @@ const routeInitialState: Map<string, any> = fromJS({
 });
 
 /**
- * Merge route into the global application state.
- */
-
-// flow-ignore-next-line (eslint & jscs bug)
-export default (state = routeInitialState, action: Object) => {
+* Merge route into the global application state.
+*/
+export default (state: Map = routeInitialState, action: Object) => {
   if (action.type === LOCATION_CHANGE) {
     return state.merge({
       locationBeforeTransitions: action.payload,
